de-duplicate email validation code
Use an utility function for validating email adresses (rather than a
full-blown service).

IMO Services should be used when the service is a object with several
method. In this case instantiating a service makes sense.

When the code is a simple function, an utility in `utils/` is simpler,
and involves less boilerplate.

This commit has the nice side-effect of actually simplifying the code
that uses the email validation: no need to inject a custom service,
or to define a private helper function.