Acceptance tests: before → beforeEach
Many acceptance tests were setting up the app once at the beginning of
the test suite (using `before`), running many tests, and destroying it
only once at the end of the whole suite.

This causes tests to be:

- Order-dependant: the app is expected to be in the state left by a previous test ;
- Fragile: the outcome of one test can leak on another.


This PR ensures that the app is re-created from scratch at the
beginning of each acceptance test case.