Result-item: remove tooltip delay
The `result-item` component currently initializes its tooltip only after
a 1s-delay. This is intended to work around an issue where the tooltip
text in the DOM is empty, although the value is correctly set on the
model.

This workaround has several issues:

- It is fragile. Random delays are never good (and this one is quite
brittle, as we will see later) ;
- It makes acceptance tests slow. The acceptance tests runner waits for
all requests, promises and timers to be resolved before executing the
next step of a test. This means every render adds 1s to the duration of
acceptance tests. Some tests take up to 3s because of this.

## The root cause

The actual issue is a sequence of event where:

1. The component is initially rendered with a `null` answer (as the
answer is not retrieved yet) ;
2. Therefore, during the initial `didRender`, the tooltip is first
initialized with a `null` tooltip ;
3. When the answer is finally available, `didRender` is called again,
and the tooltip is initialized with the correct tooltip value ;
4. Unfortunately Bootstrap’s tooltips can’t be initialized twice: they
will silently keep the initial value.

This is why the workaround _kind-of_ worked: the 1s-delay was enough to
wait for the answer being available before initializing the tooltip.
Obviously, if the answer took more than 1s to be ready, the tooltip
would break.

## The fix

The fix is simply to wait for the answer to be available before
initializing the tooltip. This behavior has been unit-tested, and should
be robust enough.

It also makes acceptance tests fast again:

- Before this PR: `time ember test` → 1m 15s
- With this PR: `time ember test` → 50s

## Caveats

We only delay the tooltip initialization for empty answers. If an
answer is set, and later changes, the tooltip will keep the initial
value.

As we don’t need this use case yet, I chose to keep the code simple and
well-tested for the cases we actually encounter. It should be easy
enough to implement updating tooltips if we ever need it.