Answer: update existing record if needed
Previously the API would return existing records on POST requests
(which should normally be used only for creating objects, not for
updating them).

Now we properly use `POST` to create new answers, and `PATCH` to
update them.

This allows to fix an Ember-Data 1.12 assertion complaining about
existing objects in the `POST` response.