Tests: fix logic of acceptance validation test
Previously the tests would visit more URLs than needed, which caused
the store to be populated with values that are not supposed to exist.

Consequently, when saving a new assessment, Ember-Data would throw an
assertion about the assessment already existing.

Fixed by making the tests only visit the required URLs.